create database once
