+++ samples/samples-csharp/Database/StoredProcedures/Drop/DropSelectProductsCost.sql
@@ -1,0 +1,1 @@
+﻿DROP PROCEDURE IF EXISTS [SelectProductsCost];
