+++ samples/samples-csharp/Database/StoredProcedures/SelectProductsCost.sql
@@ -1,6 +1,6 @@
-﻿CREATE PROCEDURE [SelectProductsCost] 
+﻿CREATE PROCEDURE [SelectProductsCost]
 	@cost INT
-AS 
-	SELECT [ProductId], [Name], [Cost] 
+AS
+	SELECT [ProductId], [Name], [Cost]
 	FROM [dbo].[Products]
 	WHERE [Cost] = @cost;
