+++ samples/samples-csharp/Database/Tables/Products.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [Products] (
+﻿DROP TABLE IF EXISTS [Products];
+
+CREATE TABLE [Products] (
 	[ProductId] [int] NOT NULL PRIMARY KEY,
 	[Name] [nvarchar](100) NULL,
 	[Cost] [int] NULL
