+++ samples/samples-csharp/Database/Tables/ProductsCostNotNull.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsCostNotNull] (
+﻿DROP TABLE IF EXISTS [ProductsCostNotNull];
+
+CREATE TABLE [ProductsCostNotNull] (
 	[ProductId] [int] NOT NULL PRIMARY KEY,
 	[Name] [nvarchar](100) NULL,
 	[Cost] [int] NOT NULL
