+++ samples/samples-csharp/Database/Tables/ProductsWithDefaultPK.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsWithDefaultPK] (
+﻿DROP TABLE IF EXISTS [ProductsWithDefaultPK];
+
+CREATE TABLE [ProductsWithDefaultPK] (
 	[ProductGuid] [uniqueidentifier] PRIMARY KEY NOT NULL DEFAULT(newsequentialid()),
 	[Name] [nvarchar](100) NULL,
 	[Cost] [int] NULL
