+++ samples/samples-csharp/Database/Tables/ProductsWithIdentity.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsWithIdentity] (
+﻿DROP TABLE IF EXISTS [ProductsWithIdentity];
+
+CREATE TABLE [ProductsWithIdentity] (
 	[ProductId] [int] PRIMARY KEY NOT NULL IDENTITY(1,1),
 	[Name] [nvarchar](100) NULL,
 	[Cost] [int] NULL
