+++ samples/samples-csharp/Database/Tables/ProductsWithMultiplePrimaryColumnsAndIdentity.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsWithMultiplePrimaryColumnsAndIdentity] (
+﻿DROP TABLE IF EXISTS [ProductsWithMultiplePrimaryColumnsAndIdentity];
+
+CREATE TABLE [ProductsWithMultiplePrimaryColumnsAndIdentity] (
 	[ProductId] [int] NOT NULL IDENTITY(1,1),
 	[ExternalId] [int] NOT NULL,
 	[Name] [nvarchar](100) NULL,
