+++ samples/samples-csharp/Database/Views/Drop/DropProductsNamesView.sql
@@ -1,0 +1,1 @@
+﻿DROP VIEW IF EXISTS ProductNames;
