+++ test/Database/Tables/ProductsUnsupportedTypes.sql
@@ -1,3 +1,5 @@
+DROP TABLE IF EXISTS [ProductsUnsupportedTypes];
+
 CREATE TABLE [ProductsUnsupportedTypes] (
     [ProductId] [int] NOT NULL PRIMARY KEY,
     [TextCol] [text],
