+++ test/Database/Tables/ProductsWithReservedPrimaryKeyColumnNames.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsWithReservedPrimaryKeyColumnNames] (
+﻿DROP TABLE IF EXISTS [ProductsWithReservedPrimaryKeyColumnNames];
+
+CREATE TABLE [ProductsWithReservedPrimaryKeyColumnNames] (
 	[ProductId] [int] NOT NULL IDENTITY(1, 1),
 	[_az_func_ChangeVersion] [int] NOT NULL,
 	[_az_func_AttemptCount] [int] NOT NULL,
