+++ test/Database/Tables/ProductsWithUnsupportedColumnTypes.sql
@@ -1,3 +1,5 @@
+DROP TABLE IF EXISTS [ProductsWithUnsupportedColumnTypes];
+
 CREATE TABLE [ProductsWithUnsupportedColumnTypes] (
     [ProductId] [int] NOT NULL PRIMARY KEY,
     [Name] [nvarchar](100) NULL,
