+++ test/Database/Tables/ProductsWithoutPrimaryKey.sql
@@ -1,4 +1,6 @@
-﻿CREATE TABLE [ProductsWithoutPrimaryKey] (
+﻿DROP TABLE IF EXISTS [ProductsWithoutPrimaryKey];
+
+CREATE TABLE [ProductsWithoutPrimaryKey] (
     [ProductId] [int] NOT NULL,
     [Name] [nvarchar](100) NULL,
     [Cost] [int] NULL
